+++ tomcat/src/main/java/org/apache/coyote/http11/FileExtractor.java
@@ -5,6 +5,7 @@ import java.io.IOException;
 import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import org.apache.coyote.http11.response.ResponseBody;
 
 public class FileExtractor {
