+++ tomcat/src/test/java/nextstep/org/apache/coyote/http11/request/HttpMethodTest.java
@@ -1,7 +1,7 @@
-package nextstep.org.apache.coyote.http11;
+package nextstep.org.apache.coyote.http11.request;
 
 import nextstep.jwp.exception.UncheckedServletException;
-import org.apache.coyote.http11.HttpMethod;
+import org.apache.coyote.http11.request.HttpMethod;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
